🚚 api: move route to src context
