+++ api/lib/application/sessions/index.js
@@ -70,19 +70,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/sessions/download-results/{token}',
-      config: {
-        auth: false,
-        handler: sessionController.getSessionResultsByRecipientEmail,
-        tags: ['api', 'sessions', 'results'],
-        notes: [
-          "Cette route est accessible via un token envoyé par email lors de l'envoi automatique des résultats de certification",
-          "Elle retourne les résultats de certifications d'une session agrégés par email de destinataire des résultats, sous format CSV",
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/sessions/download-all-results/{token}',
       config: {
         auth: false,
