+++ api/src/certification/results/application/certification-results-route.js
@@ -29,6 +29,19 @@ const register = async function (server) {
         ],
       },
     },
+    {
+      method: 'GET',
+      path: '/api/sessions/download-results/{token}',
+      config: {
+        auth: false,
+        handler: certificationResultsController.getSessionResultsByRecipientEmail,
+        tags: ['api', 'sessions', 'results'],
+        notes: [
+          "Cette route est accessible via un token envoyé par email lors de l'envoi automatique des résultats de certification",
+          "Elle retourne les résultats de certifications d'une session agrégés par email de destinataire des résultats, sous format CSV",
+        ],
+      },
+    },
   ]);
 };
